Changed vga memory pointer into a named constant
As opposed to a magic constant and following along the meatier skeleton.
In my opinion, this is one of the more interesting aspects of OS (and
driver) development. In application development, something like this -
taking a non-zero integer literal and turning it into a pointer would
surely be wrong. In OS development, though, this is correct (if not very
resilient against memory corruption).