Added initial test framework
I'm unsure of this design. And, more importantly, some of the details
about how the actual test runner will look. But it seems like a
reasonable place to start so I'll run with it for now. Also, you might
notice that test_kernel_early prints out a message that is immediately
wiped out by the kernel_early. I know. I'm thinking of having the kernel
halt on a test failure so you can see the details of the test, but
that's one of the aspects I have to decide on.