Add support for clang instead of gcc
For now I'm leaving gcc as the default, although that might change.
clang has the benefit of not requiring people to build a gcc
cross-compiler, while gcc has the benefit of not requiring people to
install clang. Hmm... Although since they have to download the gcc
source anyway, this probably isn't that much of a benefit. I'll probably
switch at some point, just not right now.