OS now prints an H when main exits
In the lower-right corner. This is meant as an exercise to demonstrate
to myself that assembly is not my strong point. That's what practice
(like this) is for, though. In any case, the real reason is that I think
that having that H there will help distinguish between an infinite loop
and a halt.