Added implementation of memcpy
Still untested. It's interesting that memcpy returns dest. I think I
agree with the stylistic suggestion that functions that mutate their
arguments should not output anything (except an error code or something
like that, perhaps). These functions are mandated by libgcc, so it's not
under my control in this case.
I also find it interesting, if unsurprising, that memcpy makes the
caller enforce non-overlapping regions.