Made memcpy support overlapping regions
This is unnecessary given memcpy's interface, but it seems reasonable to
not make things gratuitously unsafe. My next step will probably be to
pull out the common code into memmove and make memcpy be basically an
alias of memmove.
I'm curious if the memcpys of various libcs adopt my approach, or if
they behave incorrectly when the two regions overlap. I'm hard pressed
to imagine that the extra speed afforded by saving the comparison is
worth it, but I've been surprised before.