Wrote some tests for memcmp
Just the test cases, it doesn't actually run the tests yet. I'm still
stalling on devising a good framework for running tests in this
operating system. I'm trying to avoid writing or using an external
framework, but the constraints of both C and an incomplete OS make a
clean testing framework a challenge.