So I was thinking of writing a test generator...
...And this is what I imagine the tests that the generator parses to
look like.
It's not bad, although in my head it looked nicer than it does when
actually written out (it always does). I have some qualms about the
names: for instance, I'm thinking of renaming "blobs" to "arrays" and
"!blob" to "!pointer", but that's pretty minor. A bigger concern I have
is that I'm not sure the yaml parser I'm planning on using handles !foo
as I want it to in inline lists. More experimentation is required.
The test generator is just an experiment, and it might not pan out - in
which case, I'll have to write this code myself.